+++ pkg/acmedns/logging.go
@@ -8,22 +8,29 @@ import (
 
 func SetupLogging(config AcmeDnsConfig) (*zap.Logger, error) {
 	var (
-		logger *zap.Logger
-		zapCfg zap.Config
-		err    error
+		logger     *zap.Logger
+		zapCfg     zap.Config
+		err        error
+		outLogPath []string
+		errLogPath []string
 	)
 
 	logformat := "console"
 	if config.Logconfig.Format == "json" {
 		logformat = "json"
 	}
-	outputPath := "stdout"
-	if config.Logconfig.Logtype == "file" {
-		outputPath = config.Logconfig.File
+
+	if config.Logconfig.Logtype == "stdout" || config.Logconfig.Logtype == "both" {
+		outLogPath = append(outLogPath, "stdout")
+		errLogPath = append(errLogPath, "stderr")
 	}
-	errorPath := "stderr"
-	if config.Logconfig.Logtype == "file" {
-		errorPath = config.Logconfig.File
+	// I wanted to be able to specify a different file path in ErrorOutputPaths,
+	// but it seems that only internal Zap errors are output here.
+	if config.Logconfig.Logtype == "file" || config.Logconfig.Logtype == "both" {
+		if config.Logconfig.File != "" {
+			outLogPath = append(outLogPath, config.Logconfig.File)
+			errLogPath = append(errLogPath, config.Logconfig.File)
+		}
 	}
 
 	zapCfg.Level, err = zap.ParseAtomicLevel(config.Logconfig.Level)
@@ -31,8 +38,8 @@ func SetupLogging(config AcmeDnsConfig) (*zap.Logger, error) {
 		return logger, err
 	}
 	zapCfg.Encoding = logformat
-	zapCfg.OutputPaths = []string{outputPath}
-	zapCfg.ErrorOutputPaths = []string{errorPath}
+	zapCfg.OutputPaths = outLogPath
+	zapCfg.ErrorOutputPaths = errLogPath
 	zapCfg.EncoderConfig = zapcore.EncoderConfig{
 		TimeKey:     "time",
 		MessageKey:  "msg",